// src/app/shared/models/agentData.ts
export interface AgentData {
  _search: boolean;
  nd: number;
  rows: number;
  page: number;
  sidx: string;
  sord: string;
  Page: number;
  Rows: number;
  Sidx: string;
  Sord: string;
  Login: string | null;
  PersonId: number;
  PersonFio: string | null;
  Snils: string | null;
  AgentId: number;
  AgentName: string | null;
  Inn: string | null;
  Kpp: string | null;
}